GenerateMannequinCsvCommandBaseHandler
